Food Ordering App with Django
This app is like a restaurant's menu. User can see many different dishes, their ingredients and the price. User can also interact with the page (About page, different page for each dish if you click on it, etc.). Moreover, the admin of the page can edit it, but only him, changing the price, the cheff, the availability of the foot, etc. Usage of SQL too.